Add files via upload
I have scaled the parts for the OMRON TM12 and given them the correct origins. They are not connecting to the DH parameters correctly and are giving the error 'Warning: Error creating or updating Patch Error in value of property  FaceVertexCData Number of colors must equal number of vertices ', and it not loading properly.

This should be solved soon as it has improved from the previous edition.