Add files via upload
Trajectories for all dobot movements mapped out and plotted. Looks okay with some minor issues that need to be fixed. Starting positions may also need to be slightly altered.

No object or gripper as of yet and massive lag in my getRMRCTrajectory function due to the dobot plotting itself everytime this function is called. Will need a fix for this.